Changes the int type to a ushort

+++ Assets/com.initialprefabs.msdfgen/InitialPrefabs.Collections/TypeSpan.cs
@@ -3,8 +3,8 @@
         public ushort Offset;
         public ushort Size;
 
-        public static implicit operator (int offset, int size)(TypeSpan value) => (value.Offset, value.Size);
-        public static implicit operator TypeSpan((int offset, int size) v) => new TypeSpan {
+        public static implicit operator (ushort offset, ushort size)(TypeSpan value) => (value.Offset, value.Size);
+        public static implicit operator TypeSpan((ushort offset, ushort size) v) => new TypeSpan {
             Offset = (ushort)v.offset,
             Size = (ushort)v.size
         };
